Delete skipped non-dmarc-files. I got some emails to my dmarc-email-account, which did not have a zip attachment. The current code is skipping the delete/move section for such emails, so I have to log on my server and have to investigate manually. Otherwise I will get the "Skipped" notification via cron mail everytime the cron job is run.
This is not the desired behaviour, if --delete is given or an imap-move folder is set. I do want to get informed ONCE, that there is a non-dmarc-mail, however, the mail should get be deleted/moved.

This commit makes the script skip the delete/move section only, if the file could not be opened, or if there was a database error. If a non-dmarc-mail is going to be deleted, its content is printed and thus included in the cron message for furter investigations.