Use DB transaction to avoid incomplete records being added.
If one record in the report can't be processed, the entry in table
'report' will stay in the DB, as well as any other entries in
'rptrecord' from records that were processed successfully.

This leads to an misleading error

 'Skipping IMAP message with UID #5012 due to database errors.'

because the next time the IMAP client processes the message, it'll find
the entry in table 'report' from the previous incomplete processing and
issue

 'Already have <report-org> <report-id>, skipped'

and handle the report accordingly, for example move it to the IMAP folder
configured in $imapmovefolder.