Make sure the data inserted into the DB match the enum()
Extend column dkimresult to include 'unknown', as was done for spfresult
etc. in 29a9a0a8dde1f2360d36551aca5df13dc3d251d4, as it's in reports in
the wild, too. Do the same for column disposition, just in case.

Neither of DispositionType, SPFResultType nor DKIMResultType in
https://tools.ietf.org/html/rfc7489#appendix-C mention 'unknown', but
it's only a 'proposed ... schema' after all. So, no more errors like

 DBD::mysql::db do failed: Data truncated for column 'dkimresult'

The columns disposition, dkimresult and spfresult should probably be
'NOT NULL' as well, but this cannot be changed by the simple

 ALTER TABLE $table MODIFY COLUMN $col_name $col_def

used in checkDatabase(), as on finding existing NULL values this leads to

ERROR 1265 (01000): Data truncated for column 'spfresult' at row 1912